Move level design stuff into own dir
