+++ src/level_design/scene_editor.rs
@@ -1,7 +1,7 @@
 use crate::actions::{Actions, ActionsFrozen};
 use crate::file_system_interaction::game_serialization::{GameLoadRequest, GameSaveRequest};
 use crate::file_system_interaction::level_serialization::{WorldLoadRequest, WorldSaveRequest};
-use crate::spawning::{
+use crate::level_design::spawning::{
     DelayedSpawnEvent, DuplicationEvent, GameObject, ParentChangeEvent,
     SpawnEvent as SpawnRequestEvent,
 };
