+++ src/level_design/spawning/animation_link.rs
