+++ src/level_design/spawning/event.rs
@@ -1,4 +1,4 @@
-use crate::spawning::GameObject;
+use crate::level_design::spawning::GameObject;
 use bevy::prelude::*;
 use serde::{Deserialize, Serialize};
 use std::borrow::Cow;
