+++ src/level_design/spawning/objects/level.rs
@@ -1,4 +1,4 @@
-use crate::spawning::{GameObject, PrimedGameObjectSpawner};
+use crate::level_design::spawning::{GameObject, PrimedGameObjectSpawner};
 use bevy::gltf::Gltf;
 use bevy::prelude::*;
