+++ src/level_design/spawning/objects/npc.rs
@@ -1,7 +1,7 @@
+use crate::level_design::spawning::post_spawn_modification::CustomCollider;
+use crate::level_design::spawning::{GameObject, PrimedGameObjectSpawner};
 use crate::movement_gameplay::general_movement::{CharacterVelocity, Grounded, Jump, Model};
 use crate::movement_gameplay::navigation::Follower;
-use crate::spawning::post_spawn_modification::CustomCollider;
-use crate::spawning::{GameObject, PrimedGameObjectSpawner};
 use crate::world_interaction::dialog::{DialogId, DialogTarget};
 use bevy::gltf::Gltf;
 use bevy::prelude::*;
