+++ src/level_design/spawning/objects/primitives.rs
@@ -1,4 +1,4 @@
-use crate::spawning::PrimedGameObjectSpawner;
+use crate::level_design::spawning::PrimedGameObjectSpawner;
 use bevy::prelude::*;
 use bevy_rapier3d::prelude::*;
