+++ src/level_design/spawning/spawn_container.rs
@@ -1,5 +1,5 @@
-use crate::spawning::event::{DuplicationEvent, ParentChangeEvent, SpawnEvent};
-use crate::spawning::SpawnTracker;
+use crate::level_design::spawning::event::{DuplicationEvent, ParentChangeEvent, SpawnEvent};
+use crate::level_design::spawning::SpawnTracker;
 use bevy::prelude::*;
 use bevy::utils::HashMap;
 use serde::{Deserialize, Serialize};
