+++ src/lib.rs
@@ -7,15 +7,12 @@ mod actions;
 mod audio;
 mod dev;
 mod file_system_interaction;
-mod map;
+mod level_design;
 mod math;
 mod menu;
 mod movement_gameplay;
 mod physics;
-#[cfg(feature = "editor")]
-mod scene_editor;
 mod shader;
-mod spawning;
 mod trait_extension;
 mod world_interaction;
 
@@ -25,7 +22,8 @@ use crate::dev::DevPlugin;
 use crate::file_system_interaction::asset_loading::LoadingPlugin;
 use crate::file_system_interaction::game_serialization::SavingPlugin;
 use crate::file_system_interaction::level_serialization::WorldSerializationPlugin;
-use crate::map::MapPlugin;
+use crate::level_design::map::MapPlugin;
+use crate::level_design::spawning::SpawningPlugin;
 use crate::menu::MenuPlugin;
 use crate::movement_gameplay::camera::CameraPlugin;
 use crate::movement_gameplay::general_movement::MovementPlugin;
@@ -33,7 +31,6 @@ use crate::movement_gameplay::navigation::NavigationPlugin;
 use crate::movement_gameplay::player::PlayerPlugin;
 use crate::physics::PhysicsPlugin;
 use crate::shader::ShaderPlugin;
-use crate::spawning::SpawningPlugin;
 use crate::world_interaction::condition::ConditionPlugin;
 use crate::world_interaction::dialog::DialogPlugin;
 use crate::world_interaction::interactions_ui::InteractionsUi;
