+++ src/movement_gameplay/general_movement.rs
@@ -2,7 +2,7 @@ use bevy::prelude::*;
 
 use bevy_rapier3d::prelude::*;
 mod components;
-use crate::spawning::AnimationEntityLink;
+use crate::level_design::spawning::AnimationEntityLink;
 use crate::trait_extension::Vec3Ext;
 use crate::GameState;
 pub use components::*;
